remove util.dart
