+++ packages/widgetbook_cli/bin/git/bot.dart
@@ -1,6 +1,25 @@
 // Code copied from
 // https://github.com/kevmoo/bot.dart/commit/6badd135a5
 
+const shaRegexPattern = '[a-f0-9]{40}';
+
+final headerRegExp = RegExp(r'^([a-z]+) (.+)$');
+
+// ignore: prefer_interpolation_to_compose_strings
+final _shaRegEx = RegExp(r'^' + shaRegexPattern + r'$');
+
+bool isValidSha(String value) => _shaRegEx.hasMatch(value);
+
+void requireArgumentValidSha1(String value, String argName) {
+  metaRequireArgumentNotNullOrEmpty(argName);
+  requireArgumentNotNullOrEmpty(value, argName);
+
+  if (!isValidSha(value)) {
+    final message = 'Not a valid SHA1 value: $value';
+    throw ArgumentError.value(value, argName, message);
+  }
+}
+
 void requireArgument(bool truth, String argName, [String? message]) {
   metaRequireArgumentNotNullOrEmpty(argName);
   if (!truth) {
