+++ packages/widgetbook_cli/bin/git/top_level.dart
@@ -1,13 +1,6 @@
 import 'dart:async';
 import 'dart:io';
 
-import 'util.dart';
-
-// ignore: prefer_interpolation_to_compose_strings
-final _shaRegEx = RegExp(r'^' + shaRegexPattern + r'$');
-
-bool isValidSha(String value) => _shaRegEx.hasMatch(value);
-
 Future<ProcessResult> runGit(
   List<String> args, {
   bool throwOnError = true,
